Smart & dumb, partial
